pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "veri-helix"
version = "0.3.0"
description = "Friendly, reproducible bioinformatics toolkit with schema-verified viz"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{name = "Helix Authors"}]
keywords = ["bioinformatics", "genomics", "rna", "alignment", "visualization"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "Intended Audience :: Science/Research",
]
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.6",
    "networkx>=3.0",
    "imageio>=2.34",
]
protein = ["biopython>=1.84"]
schema = ["pydantic>=2.5"]
gui = ["PySide6>=6.6"]
realtime = [
    "glfw>=2.6.4",
    "moderngl>=5.8",
]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "build",
    "twine",
    "matplotlib>=3.6",
    "networkx>=3.0",
    "imageio>=2.34",
    "biopython>=1.84",
    "pydantic>=2.5",
    "psutil>=5.9",
    "PySide6>=6.6",
    "glfw>=2.6.4",
    "moderngl>=5.8",
]

[project.urls]
Homepage = "https://github.com/omniscoder/Helix"
Documentation = "https://omniscoder.github.io/Helix/"
Issues = "https://github.com/omniscoder/Helix/issues"

[project.scripts]
helix = "helix.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"helix.schema" = ["*.json"]
"helix.datasets" = ["**/*"]
"helix.gui.resources" = ["**/*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra"
